A game using Python Pygame
A game in which the user uses the left and right arrow keys to move on the screen while simultaneously avoiding the bombs, being dropped from above, and collecting the coins scattered on the ground. The user has 20 seconds to collect all 5 coins if they succeed they win if they don't finish intime they lose.